==> 02-personalize_ranking_movielens.ipynb <==
# Personalize Ranking Example <a class="anchor" id="top"></a>

In this notebook, you will choose a dataset and prepare it for use with Amazon Personalize Batch Recommendations.

1. [Choose a dataset or data source](#source)
1. [Prepare your data](#prepare)
1. [Create dataset groups and the interactions dataset](#group_dataset)
1. [Configure an S3 bucket and an IAM role](#bucket_role)
1. [Import the interactions data](#import)
1. [Create solutions](#solutions)
1. [Create campaigns](#create)
1. [Interact with campaigns](#interact)
1. [Clean up](#cleanup)

## Introduction <a class="anchor" id="intro"></a>

For the most part, the algorithms in Amazon Personalize (called recipes) look to solve different tasks, explained here:

1. **HRNN & HRNN-Metadata** - Recommends items based on previous user interactions with items.
1. **HRNN-Coldstart** - Recommends new items for which interaction data is not yet available.
1. **Personalized-Ranking** - Takes a collection of items and then orders them in probable order of interest using an HRNN-like approach.
1. **SIMS (Similar Items)** - Given one item, recommends other items also interacted with by users.
1. **Popularity-Count** - Recommends the most popular items, if HRNN or HRNN-Metadata do not have an answer - this is returned by default.

No matter the use case, the algorithms all share a base of learning on user-item-interaction data which is defined by 3 core attributes:

1. **UserID** - The user who interacted
1. **ItemID** - The item the user interacted with
1. **Timestamp** - The time at which the interaction occurred

We also support event types and event values defined by:

1. **Event Type** - Categorical label of an event (browse, purchased, rated, etc).
1. **Event Value** - A value corresponding to the event type that occurred. Generally speaking, we look for normalized values between 0 and 1 over the event types. For example, if there are three phases to complete a transaction (clicked, added-to-cart, and purchased), then there would be an event_value for each phase as 0.33, 0.66, and 1.0 respectfully.

The event type and event value fields are additional data which can be used to filter the data sent for training the personalization model. In this particular exercise we will not have an event type or event value. 

## Choose a dataset or data source <a class="anchor" id="source"></a>
[Back to top](#top)

As we mentioned, the user-item-iteraction data is key for getting started with the service. This means we need to look for use cases that generate that kind of data, a few common examples are:

1. Video-on-demand applications
1. E-commerce platforms
1. Social media aggregators / platforms

There are a few guidelines for scoping a problem suitable for Personalize. We recommend the values below as a starting point, although the [official limits](https://docs.aws.amazon.com/personalize/latest/dg/limits.html) lie a little lower.

* Authenticated users
* At least 50 unique users
* At least 100 unique items
* At least 2 dozen interactions for each user 

Most of the time this is easily attainable, and if you are low in one category, you can often make up for it by having a larger number in another category.

Generally speaking your data will not arrive in a perfect form for Personalize, and will take some modification to be structured correctly. This notebook looks to guide you through all of that. 

To begin with, we are going to use the MovieLens 100k dataset. These are records of the movie rating behavior of its users. The data fits our guidelines with a large number for users, items, and interactions.

First, you will download the dataset and unzip it in a new folder using the code below.

Take a look at the data files you have downloaded.

data_dir = "ml-100k"
!ls $data_dir

## Prepare your data <a class="anchor" id="prepare"></a>
[Back to top](#top)

The next thing to be done is to load the data and confirm the data is in a good state, then save it to a CSV where it is ready to be used with Amazon Personalize.

To get started, import a collection of Python libraries commonly used in data science.

import time
from time import sleep
import json
from datetime import datetime
import numpy as np
import boto3
import pandas as pd

%store -r rating_threshold


rating_threshold

Next,open the data file and take a look at the first several rows.

original_data = pd.read_csv(data_dir + '/train.csv',  header=0)
original_data = original_data.rename(columns={0: "user", 1: "item", 2:"rating", 3:"timestamp"})
original_data.head(5)

That's better. Now that the data has been successfully loaded into memory, let's extract some additional information. First, calculate some basic statistics from the data.

original_data.describe()

This shows that we have a good range of values for `userID` and `movieID`. Next, it is always a good idea to confirm the data format.

original_data.info()

interactions_df = original_data.copy()
interactions_df = interactions_df[['uid', 'iid', 'rating', 'timestamp']]

After manipulating the data, always confirm if the data format has changed.

 Amazon Personalize has default column names for users, items, and timestamp. These default column names are `USER_ID`, `ITEM_ID`, AND `TIMESTAMP`. So the final modification to the dataset is to replace the existing column headers with the default headers.

interactions_df.rename(columns = {'uid':'USER_ID', 'iid':'ITEM_ID', 'rating':'EVENT_VALUE',
                              'timestamp':'TIMESTAMP'}, inplace = True) 
interactions_df['EVENT_TYPE'] = "rating"

interactions_df.head()

interactions_df = interactions_df[interactions_df['EVENT_VALUE']>=rating_threshold]

That's it! At this point the data is ready to go, and we just need to save it as a CSV file.

interactions_filename = "interactions.csv"
interactions_df.to_csv((data_dir+"/"+interactions_filename), index=False, float_format='%.0f')

user_df = pd.read_csv(data_dir + '/u.user', delimiter='|', header=None, names=['USER_ID', 'AGE', 'GENDER', 'OCCUPATION', 'ZIPCODE'])


user_df['LOC'] =  user_df['ZIPCODE'].apply(lambda x:x[:3])


user_df.head()

users_filename = "users.csv"
user_df[['USER_ID','AGE','GENDER','OCCUPATION','LOC']].to_csv((data_dir+"/"+users_filename), index=False, float_format='%.0f')

genres = ['unknown','Action' , 'Adventure', 'Animation', 'Childrens' , 'Comedy' , 'Crime', \
                                        'Documentary', 'Drama' ,'Fantasy' , 'Film-Noir' , 'Horror' , 'Musical', \
                                        'Mystery', 'Romance', 'Sci-Fi', 'Thriller', 'War', 'Western']

item_df = pd.read_csv(data_dir + '/u.item', names=['ITEM_ID','title','release_date','video_release_date', 'imdb url'] + genres,  sep='|', encoding = "ISO-8859-1")


def to_timestamp(x):
    try: 
        return int(time.mktime(datetime.strptime(x, "%d-%b-%Y").timetuple())) 
    except: 
        return None
item_df['CREATION_TIMESTAMP'] = item_df['release_date'].apply(lambda x: to_timestamp(x))

item_df['CREATION_TIMESTAMP'].fillna(item_df['CREATION_TIMESTAMP'].median())



item_df[['ITEM_ID','CREATION_TIMESTAMP']+genres]

item_input_df = item_df.copy()

item_input_df['GENRE']=''

for col_name in genres:
    item_input_df.loc[item_input_df[col_name]==1,'GENRE']= item_input_df['GENRE']+'|'+ col_name

item_input_df = item_input_df[['ITEM_ID','CREATION_TIMESTAMP', 'GENRE']]
item_input_df

items_filename = "items.csv"
item_input_df.to_csv((data_dir+"/"+items_filename), index=False, float_format='%.0f')

## Create dataset groups and the interactions dataset <a class="anchor" id="group_dataset"></a>
[Back to top](#top)

The highest level of isolation and abstraction with Amazon Personalize is a *dataset group*. Information stored within one of these dataset groups has no impact on any other dataset group or models created from one - they are completely isolated. This allows you to run many experiments and is part of how we keep your models private and fully trained only on your data. 

Before importing the data prepared earlier, there needs to be a dataset group and a dataset added to it that handles the interactions.

Dataset groups can house the following types of information:

* User-item-interactions
* Event streams (real-time interactions)
* User metadata
* Item metadata

Before we create the dataset group and the dataset for our interaction data, let's validate that your environment can communicate successfully with Amazon Personalize.

# Configure the SDK to Personalize:
personalize = boto3.client('personalize')
personalize_runtime = boto3.client('personalize-runtime')

### Create the dataset group

The following cell will create a new dataset group with specified name 

Before we can use the dataset group, it must be active. This can take a minute or two. Execute the cell below and wait for it to show the ACTIVE status. It checks the status of the dataset group every second, up to a maximum of 3 hours.

def create_dataset_group(name): 
    create_dataset_group_response = personalize.create_dataset_group(
        name = name
    )

    dataset_group_arn = create_dataset_group_response['datasetGroupArn']
    print(json.dumps(create_dataset_group_response, indent=2))
    max_time = time.time() + 3*60*60 # 3 hours
    while time.time() < max_time:
        describe_dataset_group_response = personalize.describe_dataset_group(
            datasetGroupArn = dataset_group_arn
        )
        status = describe_dataset_group_response["datasetGroup"]["status"]
        print("DatasetGroup: {}".format(status))
    
        if status == "ACTIVE" or status == "CREATE FAILED":
            break
        
        time.sleep(60)
    return dataset_group_arn

prj_suffix = str(np.random.uniform())[4:9]
dataset_group_name = "p13n-ml100k-{}".format(prj_suffix)

dataset_group_arn = create_dataset_group(dataset_group_name)

Now that you have a dataset group, you can create a dataset for the interaction data.

### Create the dataset

First, define a schema to tell Amazon Personalize what type of dataset you are uploading. There are several reserved and mandatory keywords required in the schema, based on the type of dataset. More detailed information can be found in the [documentation](https://docs.aws.amazon.com/personalize/latest/dg/how-it-works-dataset-schema.html).

Here, you will create a schema for interactions data, which needs the `USER_ID`, `ITEM_ID`, and `TIMESTAMP` fields. These must be defined in the same order in the schema as they appear in the dataset.

interactions_schema = {
    "type": "record",
    "name": "Interactions",
    "namespace": "com.amazonaws.personalize.schema",
    "fields": [
        {
            "name": "USER_ID",
            "type": "string"
        },
        {
            "name": "ITEM_ID",
            "type": "string"
        },
        {
          "name": "EVENT_VALUE",
          "type": [
             "float",
             "null"
          ]
        },
        {
            "name": "TIMESTAMP",
            "type": "long"
        },
        {
            "name": "EVENT_TYPE",
            "type": "string"
        },
    ],
    "version": "1.0"
}

create_schema_response = personalize.create_schema(
    name = "p13n-ml100k-interaction-{}".format(prj_suffix),
    schema = json.dumps(interactions_schema)
)

schema_arn = create_schema_response['schemaArn']
print(json.dumps(create_schema_response, indent=2))

With a schema created, you can create a dataset within the dataset group. Note, this does not load the data yet. This will happen a few steps later.

dataset_type = "INTERACTIONS"
create_dataset_response = personalize.create_dataset(
    name = "p13n-ml100k-interaction",
    datasetType = dataset_type,
    datasetGroupArn = dataset_group_arn,
    schemaArn = schema_arn
)

interactions_dataset_arn = create_dataset_response['datasetArn']
print(json.dumps(create_dataset_response, indent=2))

def add_genres(fields, genres): 
    for g in genres: 
        genre_attr = {
          "name": g ,
          "type": [
            "null",
            "int"
          ],
          "categorical": True
        }
    fields.append(genre_attr)
    
    
fields = [
    {
      "name": "ITEM_ID",
      "type": "string"
    },
    {
      "name": "CREATION_TIMESTAMP",
      "type": "long"
}]


add_genres(fields, genres)

items_schema = {
  "type": "record",
  "name": "Items",
  "namespace": "com.amazonaws.personalize.schema",
  "fields": fields,  
  "version": "1.0"
}

create_schema_response = personalize.create_schema(
    name = "p13n-ml100k-items-{}".format(prj_suffix),
    schema = json.dumps(items_schema)
)

item_schema_arn = create_schema_response['schemaArn']



# add_genres(fields, genres)

items_schema = {
  "type": "record",
  "name": "Items",
  "namespace": "com.amazonaws.personalize.schema",
  "fields":  [
    {
      "name": "ITEM_ID",
      "type": "string"
    },
    {
      "name": "CREATION_TIMESTAMP",
      "type": "long"
    }, 
    {
        "name": "GENRE",
        "type": [
            "null",
            "string"
        ],
        "categorical": True
    }
  ],  
  "version": "1.0"
}

create_schema_response = personalize.create_schema(
    name = "p13n-ml100k-items-{}".format(prj_suffix),
    schema = json.dumps(items_schema)
)

item_schema_arn = create_schema_response['schemaArn']


dataset_type = "ITEMS"
create_dataset_response = personalize.create_dataset(
    name = "p13n-ml100k-items",
    datasetType = dataset_type,
    datasetGroupArn = dataset_group_arn,
    schemaArn = item_schema_arn
)

items_dataset_arn = create_dataset_response['datasetArn']
print(json.dumps(create_dataset_response, indent=2))


# USER_ID	AGE	GENDER	OCCUPATION	ZIPCODE

user_schema = {
  "type": "record",
  "name": "Users",
  "namespace": "com.amazonaws.personalize.schema",
  "fields": [
    {
      "name": "USER_ID",
      "type": "string"
    },
    {
      "name": "AGE",
      "type": [
        "null",
        "int"
      ],
      "categorical": False
    },
    {
      "name": "GENDER",
      "type": [
        "null",
        "string"
      ],
      "categorical": True
    },
       {
      "name": "OCCUPATION",
      "type": [
        "null",
        "string"
      ],
      "categorical": True
    },
    {
      "name": "LOC",
      "type": [
        "null",
        "string"
      ],
      "categorical": True
    },

  ],
  "version": "1.0"
}

create_schema_response = personalize.create_schema(
    name = "p13n-ml100k-users-{}".format(prj_suffix),
    schema = json.dumps(user_schema)
)

user_schema_arn = create_schema_response['schemaArn']


# personalize.delete_dataset(
#     datasetArn = users_dataset_arn
# )

# response = personalize.delete_schema(
#     schemaArn='arn:aws:personalize:us-west-2:230755935769:schema/p13n-ml100k-users-20211127'
# )


dataset_type = "USERS"
create_dataset_response = personalize.create_dataset(
    name = "p13n-ranking-ml100k-users",
    datasetType = dataset_type,
    datasetGroupArn = dataset_group_arn,
    schemaArn = user_schema_arn
)

users_dataset_arn = create_dataset_response['datasetArn']
print(json.dumps(create_dataset_response, indent=2))

## Configure an S3 bucket and an IAM  role <a class="anchor" id="bucket_role"></a>
[Back to top](#top)

So far, we have downloaded, manipulated, and saved the data onto the Amazon EBS instance attached to instance running this Jupyter notebook. However, Amazon Personalize will need an S3 bucket to act as the source of your data, as well as IAM roles for accessing that bucket. Let's set all of that up.

Use the metadata stored on the instance underlying this Amazon SageMaker notebook, to determine the region it is operating in. If you are using a Jupyter notebook outside of Amazon SageMaker, simply define the region as a string below. The Amazon S3 bucket needs to be in the same region as the Amazon Personalize resources we have been creating so far.

with open('/opt/ml/metadata/resource-metadata.json') as notebook_info:
    data = json.load(notebook_info)
    resource_arn = data['ResourceArn']
    region = resource_arn.split(':')[3]
print(region)

Amazon S3 bucket names are globally unique. To create a unique bucket name, the code below will append the string `personalizepoc` to your AWS account number. Then it creates a bucket with this name in the region discovered in the previous cell.

s3 = boto3.client('s3')

bucket_name = "p13n-movielens-ranking-demo-"+   prj_suffix        # replace with the name of your S3 bucket
print(bucket_name)
if region != "us-east-1":
    s3.create_bucket(Bucket=bucket_name, CreateBucketConfiguration={'LocationConstraint': region})
else:
    s3.create_bucket(Bucket=bucket_name)

### Upload data to S3

Now that your Amazon S3 bucket has been created, upload the CSV file of our user-item-interaction data. 

interactions_file_path = data_dir + "/" + interactions_filename
boto3.Session().resource('s3').Bucket(bucket_name).Object(interactions_filename).upload_file(interactions_file_path)
interactions_s3DataPath = "s3://"+bucket_name+"/"+interactions_filename
interactions_s3DataPath

users_file_path = data_dir + "/" + users_filename
boto3.Session().resource('s3').Bucket(bucket_name).Object(users_filename).upload_file(users_file_path)
users_s3DataPath = "s3://"+bucket_name+"/"+users_filename
users_s3DataPath

items_file_path = data_dir + "/" + items_filename
boto3.Session().resource('s3').Bucket(bucket_name).Object(items_filename).upload_file(items_file_path)
items_s3DataPath = "s3://"+bucket_name+"/"+items_filename
items_s3DataPath

### Set the S3 bucket policy
Amazon Personalize needs to be able to read the contents of your S3 bucket. So add a bucket policy which allows that.

policy = {
    "Version": "2012-10-17",
    "Id": "PersonalizeS3BucketAccessPolicy",
    "Statement": [
        {
            "Sid": "PersonalizeS3BucketAccessPolicy",
            "Effect": "Allow",
            "Principal": {
                "Service": "personalize.amazonaws.com"
            },
            "Action": [
                "s3:*Object",
                "s3:ListBucket"
            ],
            "Resource": [
                "arn:aws:s3:::{}".format(bucket_name),
                "arn:aws:s3:::{}/*".format(bucket_name)
            ]
        }
    ]
}

s3.put_bucket_policy(Bucket=bucket_name, Policy=json.dumps(policy))

### Create an IAM role

Amazon Personalize needs the ability to assume roles in AWS in order to have the permissions to execute certain tasks. Let's create an IAM role and attach the required policies to it. The code below attaches very permissive policies; please use more restrictive policies for any production application.

iam = boto3.client("iam")

role_name = "MovieLens100kP13NRole-{}".format(prj_suffix)
assume_role_policy_document = {
    "Version": "2012-10-17",
    "Statement": [
        {
          "Effect": "Allow",
          "Principal": {
            "Service": "personalize.amazonaws.com"
          },
          "Action": "sts:AssumeRole"
        }
    ]
}

create_role_response = iam.create_role(
    RoleName = role_name,
    AssumeRolePolicyDocument = json.dumps(assume_role_policy_document)
)

# AmazonPersonalizeFullAccess provides access to any S3 bucket with a name that includes "personalize" or "Personalize" 
# if you would like to use a bucket with a different name, please consider creating and attaching a new policy
# that provides read access to your bucket or attaching the AmazonS3ReadOnlyAccess policy to the role
policy_arn = "arn:aws:iam::aws:policy/service-role/AmazonPersonalizeFullAccess"
iam.attach_role_policy(
    RoleName = role_name,
    PolicyArn = policy_arn
)

# Now add S3 support
iam.attach_role_policy(
    PolicyArn='arn:aws:iam::aws:policy/AmazonS3FullAccess',
    RoleName=role_name
)
time.sleep(60) # wait for a minute to allow IAM role policy attachment to propagate

role_arn = create_role_response["Role"]["Arn"]
print(role_arn)

## Import the interactions data <a class="anchor" id="import"></a>
[Back to top](#top)

Earlier you created the dataset group and dataset to house your information, so now you will execute an import job that will load the data from the S3 bucket into the Amazon Personalize dataset. 

Before we can use the dataset, the import job must be active. Execute the cell below and wait for it to show the ACTIVE status. It checks the status of the import job every second, up to a maximum of 3 hours.

Importing the data can take some time, depending on the size of the dataset. In this workshop, the data import job should take around 15 minutes.

def create_import_job(job_name, dataset_arn, s3DataPath, role_arn): 
    create_dataset_import_job_response = personalize.create_dataset_import_job(
        jobName = job_name,
        datasetArn = dataset_arn,
        dataSource = {
            "dataLocation": s3DataPath
        },
        roleArn = role_arn
    )

    dataset_import_job_arn = create_dataset_import_job_response['datasetImportJobArn']
    print(json.dumps(create_dataset_import_job_response, indent=2))
    
    max_time = time.time() + 3*60*60 # 3 hours
    while time.time() < max_time:
        describe_dataset_import_job_response = personalize.describe_dataset_import_job(
            datasetImportJobArn = dataset_import_job_arn
        )
        status = describe_dataset_import_job_response["datasetImportJob"]['status']
        print("DatasetImportJob: {}".format(status))

        if status == "ACTIVE" or status == "CREATE FAILED":
            break

        time.sleep(60)
    

create_import_job("p13n-ml100k-ranking-ui-job-{}".format(prj_suffix),interactions_dataset_arn, interactions_s3DataPath, role_arn)

create_import_job("p13n-ml100k-ranking-items-job-{}".format(prj_suffix),items_dataset_arn, items_s3DataPath, role_arn)

create_import_job("p13n-ml100k-ranking-users-job-{}".format(prj_suffix),users_dataset_arn, users_s3DataPath, role_arn)

When the dataset import is active, you are ready to start building models with SIMS, Personalized-Ranking, Popularity-Count, and HRNN. This process will continue in other notebooks. Run the cell below before moving on to store a few values for usage in the next notebooks.

## Create solutions <a class="anchor" id="solutions"></a>
[Back to top](#top)

In this notebook, you will create solutions with the following recipe:

`aws-personalized-ranking`


In Amazon Personalize, a specific variation of an algorithm is called a recipe. Different recipes are suitable for different situations. A trained model is called a solution, and each solution can have many versions that relate to a given volume of data when the model was trained.

To start, we will list all the recipes that are supported. This will allow you to select one and use that to build your model.

personalize.list_recipes()

### Use cases - 

#### User personalized recommendation - Usually on home page or EDM 
#### Similar items - Usually on item page to increase user engagement
#### Personalized ranking - rerank a list of items based on the user's preference 

p13n_recipe_arn="arn:aws:personalize:::recipe/aws-user-personalization"
rerank_recipe_arn = "arn:aws:personalize:::recipe/aws-personalized-ranking"
sims_recipe_arn = "arn:aws:personalize:::recipe/aws-similar-items"

#### Create the solution

As with the previous solution, start by creating the solution first. Although you provide the dataset ARN in this step, the model is not yet trained. See this as an identifier instead of a trained model.

#### Create the solution version

Once you have a solution, you need to create a version in order to complete the model training. The training can take a while to complete, upwards of 25 minutes, and an average of 35 minutes for this recipe with our dataset. Normally, we would use a while loop to poll until the task is completed. However the task would block other cells from executing, and the goal here is to create many models and deploy them quickly. So we will set up the while loop for all of the solutions further down in the notebook. There, you will also find instructions for viewing the progress in the AWS console.

#### View solution creation status

As promised, how to view the status updates in the console:

* In another browser tab you should already have the AWS Console up from opening this notebook instance. 
* Switch to that tab and search at the top for the service `Personalize`, then go to that service page. 
* Click `View dataset groups`.
* Click the name of your dataset group, most likely something with POC in the name.
* Click `Solutions and recipes`.
* You will now see a list of all of the solutions you created above,  including a column with the status of the solution versions. Once it is `Active`, your solution is ready to be reviewed. It is also capable of being deployed.

Or simply run the cell below to keep track of the solution version creation status.






def create_solution(solution_name, dataset_group_arn, recipe, hpo=False, solution_config=None, wait=True): 
    
    if hpo and solution_config: 
        create_solution_response = personalize.create_solution(
            name = solution_name,
            datasetGroupArn = dataset_group_arn,
            recipeArn = recipe,
            performHPO = hpo, 
            solutionConfig = solution_config
        )
    else:     
        create_solution_response = personalize.create_solution(
            name = solution_name,
            datasetGroupArn = dataset_group_arn,
            recipeArn = recipe
        )

    solution_arn = create_solution_response['solutionArn']
    print(json.dumps(create_solution_response, indent=2))
    create_solution_version_response = personalize.create_solution_version(
        solutionArn = solution_arn
    )
    solution_version_arn = create_solution_version_response['solutionVersionArn']
    in_progress_solution_versions = [
        solution_version_arn
    ]
    if not wait: 
        return solution_version_arn
    max_time = time.time() + 3*60*60 # 3 hours
    while time.time() < max_time:
        for solution_version_arn in in_progress_solution_versions:
            version_response = personalize.describe_solution_version(
                solutionVersionArn = solution_version_arn
            )
            status = version_response["solutionVersion"]["status"]

            if status == "ACTIVE":
                print("Build succeeded for {}".format(solution_version_arn))
                in_progress_solution_versions.remove(solution_version_arn)
            elif status == "CREATE FAILED":
                print("Build failed for {}".format(solution_version_arn))
                in_progress_solution_versions.remove(solution_version_arn)

        if len(in_progress_solution_versions) <= 0:
            break
        else:
            print("At least one solution build is still in progress")

        time.sleep(60)
    return solution_version_arn 


solution_config =  {
      "algorithmHyperParameters": {
          "hidden_dimension": "55"
      },
      "hpoConfig": {
          "algorithmHyperParameterRanges": {
              "categoricalHyperParameterRanges": [
                  {
                      "name": "recency_mask",
                      "values": [ "true", "false" ]
                  }
              ],
              
              'continuousHyperParameterRanges': [
              ],
              
              "integerHyperParameterRanges": [
                  {
                      "name": "bptt",
                      "minValue": 2,
                      "maxValue": 32
                  }
              ]
          },
          "hpoResourceConfig": {
              "maxNumberOfTrainingJobs": "4",
              "maxParallelTrainingJobs": "2"
          }
      }
  }
p13n_solution_version_arn = create_solution("p13n-ml100k-{}".format(prj_suffix), dataset_group_arn, p13n_recipe_arn, True, solution_config)

p13n_solution_version_arn = create_solution("p13n-ml100k-{}".format(prj_suffix), dataset_group_arn, p13n_recipe_arn, True, solution_config)

## Create campaigns <a class="anchor" id="create"></a>
[Back to top](#top)

A campaign is a hosted solution version; an endpoint which you can query for recommendations. Pricing is set by estimating throughput capacity (requests from users for personalization per second). When deploying a campaign, you set a minimum throughput per second (TPS) value. This service, like many within AWS, will automatically scale based on demand, but if latency is critical, you may want to provision ahead for larger demand. For this POC and demo, all minimum throughput thresholds are set to 1. For more information, see the [pricing page](https://aws.amazon.com/personalize/pricing/).

Let's start deploying the campaigns.

### User Personalized Recommendation Campaign 

Deploy a campaign for your personalized ranking solution version. It can take around 10 minutes to deploy a campaign. Normally, we would use a while loop to poll until the task is completed. However the task would block other cells from executing, and the goal here is to create multiple campaigns. So we will set up the while loop for all of the campaigns further down in the notebook. There, you will also find instructions for viewing the progress in the AWS console.

### View campaign creation status

As promised, how to view the status updates in the console:

* In another browser tab you should already have the AWS Console up from opening this notebook instance. 
* Switch to that tab and search at the top for the service `Personalize`, then go to that service page. 
* Click `View dataset groups`.
* Click the name of your dataset group, most likely something with POC in the name.
* Click `Campaigns`.
* You will now see a list of all of the campaigns you created above, including a column with the status of the campaign. Once it is `Active`, your campaign is ready to be queried.

Or simply run the cell below to keep track of the campaign creation status.

def create_campaign(campaign_name, solution_version_arn):
    create_campaign_response = personalize.create_campaign(
        name = campaign_name,
        solutionVersionArn = solution_version_arn,
        minProvisionedTPS = 1
    )

    campaign_arn = create_campaign_response['campaignArn']
    in_progress_campaigns = [
        campaign_arn
    ]

    max_time = time.time() + 3*60*60 # 3 hours
    while time.time() < max_time:
        for campaign_arn in in_progress_campaigns:
            version_response = personalize.describe_campaign(
                campaignArn = campaign_arn
            )
            status = version_response["campaign"]["status"]

            if status == "ACTIVE":
                print("Build succeeded for {}".format(campaign_arn))
                in_progress_campaigns.remove(campaign_arn)
            elif status == "CREATE FAILED":
                print("Build failed for {}".format(campaign_arn))
                in_progress_campaigns.remove(campaign_arn)

        if len(in_progress_campaigns) <= 0:
            break
        else:
            print("At least one campaign build is still in progress")

        time.sleep(60)
    return campaign_arn 

p13n_campaign = create_campaign("p13n-ml100k-{}".format(prj_suffix), p13n_solution_version_arn)

## Interact with campaigns <a class="anchor" id="interact"></a>
[Back to top](#top)

Now that all campaigns are deployed and active, we can start to get recommendations via an API call. Each of the campaigns is based on a different recipe, which behave in slightly different ways because they serve different use cases. We will cover each campaign in a different order than used in previous notebooks, in order to deal with the possible complexities in ascending order (i.e. simplest first).

First, let's create a supporting function to help make sense of the results returned by a Personalize campaign. Personalize returns only an `item_id`. This is great for keeping data compact, but it means you need to query a database or lookup table to get a human-readable result for the notebooks. We will create a helper function to return a human-readable result from the MovieLens dataset.

Start by loading in the dataset which we can use for our lookup table.

items_df = pd.read_pickle('item_df.p')

items_df

def get_movie_by_id(movie_id, movie_df=items_df):
    """
    This takes in an movie_id from Personalize so it will be a string,
    converts it to an int, and then does a lookup in a default or specified
    dataframe.
    
    A really broad try/except clause was added in case anything goes wrong.
    
    Feel free to add more debugging or filtering here to improve results if
    you hit an error.
    """
    try:
        c_row = movie_df[movie_df['iid']==movie_id].iloc[0]
        title = c_row['title'] 
        m_genres = [] 
        for g in genres: 
            if c_row[g] == 1: 
                m_genres.append(g)
        return movie_df[movie_df['iid']==movie_id].iloc[0]['title'] + " genres:" + ",".join(m_genres)
    except:
        return "Error obtaining movie info"

Now let's test a few simple values to check our error catching.

# A known good id
print(get_movie_by_id(movie_id=1))

Great! Now we have a way of rendering results. 

### Personalized Ranking

The core use case for personalized ranking is to take a collection of items and to render them in priority or probable order of interest for a user. To demonstrate this, we will need a random user and a random collection of 25 items.

Then make the personalized ranking API call.

user_df = pd.read_pickle('user_df.p')

user_df

import matplotlib.pyplot as plt

user_id = str(1)
# plt.pie(user_df[user_df['uid']==1].values[0][3:22], labels=genres)
plt.pie(user_df[user_df['uid']==int(user_id)].values[0][2:21], labels=genres)

import json

import numpy 
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd


user_item_df = pd.read_pickle("user_item_df.p")
item_df = pd.read_pickle("item_df.p")
genres = ['unknown','Action' , 'Adventure', 'Animation', 'Childrens' , 'Comedy' , 'Crime', \
                                        'Documentary', 'Drama' ,'Fantasy' , 'Film-Noir' , 'Horror' , 'Musical', \
                                        'Mystery', 'Romance', 'Sci-Fi', 'Thriller', 'War', 'Western']


def plot_heat_map(df, figsize=(10,7)): 
    df = df.div(df.sum(axis=1), axis=0)     
    plt.subplots(figsize=figsize)
    sns.heatmap(df)


u_id = user_id

tester_df = user_item_df[user_item_df['uid']==int(u_id)]
tester_df['positive'] = tester_df['rating'] >3 

review = tester_df[['positive']+genres].groupby(['positive']).sum()
plot_heat_map(review, figsize=(10,5))


# p13n_campaign="arn:aws:personalize:us-west-2:230755935769:campaign/p13n-ml100k-09994"

# Get recommended reranking
get_recommendations_response = personalize_runtime.get_recommendations(
        campaignArn = p13n_campaign,
        userId = user_id,
)

get_recommendations_response

Now add the reranked items as a second column to the original dataframe, for a side-by-side comparison.

ranked_list = []
item_list = get_recommendations_response['itemList']
for item in item_list:
    movie = get_movie_by_id(int(item['itemId']))
    ranked_list.append(movie)
ranked_df = pd.DataFrame(ranked_list, columns = ['Re-Ranked'])
# compare_df = pd.concat([rerank_df, ranked_df], axis=1)
pd.set_option('display.max_colwidth', -1)
ranked_df

You can see above how each entry was re-ordered based on the model's understanding of the user. This is a popular task when you have a collection of items to surface a user, a list of promotions for example, or if you are filtering on a category and want to show the most likely good items.

### Similar Items by Amazon Personalize Recipe 

sims_solution_version_arn = create_solution("p13n-ml100k-sims-{}".format(prj_suffix), dataset_group_arn, sims_recipe_arn)
# ranking_solution_version_arn = create_solution("p13n-ml100k-ranking-{}".format(prj_suffix), dataset_group_arn, rerank_recipe_arn)

sims_campaign = create_campaign("p13n-ml100k-sims-{}".format(prj_suffix), sims_solution_version_arn)

### Personalized ranking by Amazon Personalize Recipe 

ranking_solution_version_arn = create_solution("p13n-ml100k-ranking-{}".format(prj_suffix), dataset_group_arn, rerank_recipe_arn)

ranking_campaign = create_campaign("p13n-ml100k-ranking-{}".format(prj_suffix), ranking_solution_version_arn)

item_id = 987
print(get_movie_by_id(movie_id=item_id))

sims_campaign 
get_sims_items_response = personalize_runtime.get_recommendations(
        campaignArn = sims_campaign,
        itemId = str(item_id),
)

get_sims_items_response

sim_list = []
sim_ids = []
for item in get_sims_items_response['itemList']:
    item_id = item['itemId']
    sim_ids.append(item_id)
    movie = get_movie_by_id(int(item_id)) 
    sim_list.append(movie)
sim_df = pd.DataFrame(sim_list, columns = ['original'])
sim_df


get_recommendations_response_rerank = personalize_runtime.get_personalized_ranking(
        campaignArn = ranking_campaign,
#         userId = user_id,
        userId=str(2),
        inputList = sim_ids
)

get_recommendations_response_rerank

ranked_list = []
item_list = get_recommendations_response_rerank['personalizedRanking']
for item in item_list:
    movie = get_movie_by_id(int(item['itemId']))
    ranked_list.append(movie)
ranked_df = pd.DataFrame(ranked_list, columns = ['Re-Ranked'])
compare_df = pd.concat([sim_df, ranked_df], axis=1)
pd.set_option('display.max_colwidth', -1)
compare_df
